// src/app/models/comodo.ts
export class Comodo {
    id: string;
    userID: string;
    nomeComodo: string;
    tempComodo: string;
    umiComodoF: string;
    ajustTempComodo: string;
    statusLuzComodo: boolean;
    statusTomadaComodo: boolean;
    statusJanelaComodo: boolean;
    statusPortaComodo: boolean;
    statusPresencaComodo: boolean;
    statusArCondiconado: boolean;
    isActive: boolean;
    isFavorite: boolean;
    issync: boolean;
    date: Date;
}